feat(gh-init): detect owner, deploy docs via mkdocs, and split into a package
- detect_owner: resolve the GitHub owner from pyproject [project.urls]
  (newly added to both templates), falling back to cog.toml's
  [changelog].owner; the resolved source is shown in the summary. Fixes
  repos being created under the authenticated user instead of the org.
- Pages: gh-init now runs the project's deploy-gh-pages just recipe
  (mkdocs gh-deploy --force) so gh-pages ships the real Material site
  immediately, then enables GitHub Pages and sets the repo Website.
  Skipped when there is no mkdocs.yml; failures are non-fatal.
- Workspace mkdocs: mkdocstrings paths is now packages/*/src so every
  workspace member gets a reference page, not just the first.
- Split the monolithic github_init.py into a package: config.py
  (dataclasses + resolution), client.py (PyGithub wrapper), and
  __init__.py (orchestration + git/docs subprocess helpers). Public API
  re-exported so cli/tests are unchanged.
