+++ repo_scaffold/templates/template-python/{{cookiecutter.project_slug}}/pyproject.toml
@@ -12,6 +12,11 @@ dependencies = [
 {%- endif %}
 ]
 
+[project.urls]
+Homepage = "https://github.com/{{cookiecutter.github_username}}/{{cookiecutter.project_slug}}"
+Repository = "https://github.com/{{cookiecutter.github_username}}/{{cookiecutter.project_slug}}"
+Documentation = "https://{{cookiecutter.github_username}}.github.io/{{cookiecutter.project_slug}}"
+
 [project.optional-dependencies]
 dev = [
     "pytest>=8.3.4",
